+++ pasqal_cloud/utils/filters.py
@@ -7,7 +7,17 @@ from pydantic import BaseModel, Field, field_serializer, field_validator
 from pasqal_cloud.utils.constants import JobStatus
 
 
-class BaseFilters(BaseModel):
+class BaseJobFilters(BaseModel):
+    """
+    Base class used by Job related filters for shared fields.
+    """
+
+    id: Optional[Union[List[Union[UUID, str]], Union[UUID, str]]] = None
+    min_runs: Optional[int] = None
+    max_runs: Optional[int] = None
+    start_date: Optional[datetime] = None
+    end_date: Optional[datetime] = None
+
     @staticmethod
     def convert_to_list(value: Any) -> Any:
         if not value:
@@ -30,7 +40,7 @@ class BaseFilters(BaseModel):
         return date.isoformat()
 
 
-class RebatchFilters(BaseFilters):
+class RebatchFilters(BaseJobFilters):
     """
     Class to provide filters for querying jobs to re-create.
 
@@ -52,12 +62,7 @@ class RebatchFilters(BaseFilters):
         end_date: Retry jobs created at or before this datetime.
     """
 
-    id: Optional[Union[List[Union[UUID, str]], Union[UUID, str]]] = None
     status: Optional[Union[List[JobStatus], JobStatus]] = None
-    min_runs: Optional[int] = None
-    max_runs: Optional[int] = None
-    start_date: Optional[datetime] = None
-    end_date: Optional[datetime] = None
 
     @field_validator("id", "status", mode="before")
     @classmethod
@@ -76,7 +81,41 @@ class RebatchFilters(BaseFilters):
         return [job_status.value for job_status in job_statuses]
 
 
-class JobFilters(BaseFilters):
+class CancelJobFilters(BaseJobFilters):
+    """
+    Class to provide filters for cancelling a group of jobs.
+
+    Setting a value for any attribute of this class will add a filter to the query.
+
+    When using several filters at the same time, the API will return elements who pass
+    all filters at the same time:
+        - If the filter value is a single element, the API will return jobs whose
+          attribute matches the provided value.
+        - If the filter value is a list, the API will return jobs whose value for
+          this attribute is contained in that list.
+
+    Attributes:
+        id: Filter by job IDs
+        min_runs: Minimum number of runs.
+        max_runs: Maximum number of runs.
+        start_date: Retry jobs created at or after this datetime.
+        end_date: Retry jobs created at or before this datetime.
+    """
+
+    @field_validator("id", mode="before")
+    @classmethod
+    def single_item_to_list_validator(cls, values: Dict[str, Any]) -> Any:
+        return cls.convert_to_list(values)
+
+    @field_validator("id", mode="after")
+    @classmethod
+    def str_to_uuid_validator(cls, values: Dict[str, Any]) -> Dict[str, UUID]:
+        for item in values:
+            cls.convert_str_to_uuid(item)
+        return values
+
+
+class JobFilters(BaseJobFilters):
     """
     Class to provide filters for querying jobs.
 
@@ -102,16 +141,11 @@ class JobFilters(BaseFilters):
         end_date: Retry jobs created at or before this datetime.
     """
 
-    id: Optional[Union[List[Union[UUID, str]], Union[UUID, str]]] = None
     project_id: Optional[Union[List[Union[UUID, str]], Union[UUID, str]]] = None
     user_id: Optional[Union[List[str], str]] = None
     batch_id: Optional[Union[List[Union[UUID, str]], Union[UUID, str]]] = None
     status: Optional[Union[List[JobStatus], JobStatus]] = None
-    min_runs: Optional[int] = None
-    max_runs: Optional[int] = None
     errors: Optional[bool] = None
-    start_date: Optional[datetime] = None
-    end_date: Optional[datetime] = None
 
     @field_validator("id", "project_id", "user_id", "batch_id", "status", mode="before")
     @classmethod
@@ -130,7 +164,7 @@ class JobFilters(BaseFilters):
         return [job_status.value for job_status in job_statuses]
 
 
-class PaginationParams(BaseFilters):
+class PaginationParams(BaseModel):
     """
     Class providing parameters for paginating query results.
 
@@ -142,20 +176,3 @@ class PaginationParams(BaseFilters):
 
     offset: int = Field(default=0, strict=True, ge=0)
     limit: int = Field(default=100, strict=True, gt=0, le=100)
-
-
-class PaginatedResponse(BaseModel):
-    """
-    Class representing a paginated response structure.
-
-    Attributes:
-        total: The total number of items matching the query.
-        offset: The starting point for the current set of paginated results.
-                It indicates the number of items skipped before the current set.
-        results: A list of items that match the query and pagination parameters
-                 provided.
-    """
-
-    total: int
-    offset: int
-    results: List[Any]